Finished initial overhaul to material design
Need a proper fixed-width font similar to Courier New or Consolas or FreeMono to embed.

Need to fix the listbox, kind of weird atm.

Need to make text boxes bitter, and overhaul that style entirely.

Added dependency to a _fork_ of gwt-material-design because by default it defines too many styles I don't want. Also that project appears to be under heavy development so I'm thinking I'll stick to the fork at least in the near-term.

Anyway, as part of this deal, the project now also has an (albeit indirect) dependency to JQuery :(